Wrote one test and took far too long over it...

+++ Darts/src/burlton/dartzee/code/db/DartsMatchEntity.kt
@@ -125,8 +125,8 @@ class DartsMatchEntity : AbstractEntity<DartsMatchEntity>()
         sb.append(" AND DtFinish < ")
         sb.append(getEndOfTimeSqlString())
 
-        val count = DatabaseUtil.executeQueryAggregate(sb).toLong()
-        return count == games.toLong()
+        val count = DatabaseUtil.executeQueryAggregate(sb)
+        return count == games
     }
 
     fun getPlayerCount(): Int = players.size
